Rename src/gatsby/ files
